Migrate the interaction functions with the endpoint `/account-recovery/lookup` from the old `redux thunk` to the data layer.
