+++ client/state/account-recovery/reset/actions.js
@@ -1,7 +1,6 @@
 /**
  * Internal dependencies
  */
-import wpcom from 'lib/wp';
 import {
 	ACCOUNT_RECOVERY_RESET_OPTIONS_REQUEST,
 	ACCOUNT_RECOVERY_RESET_OPTIONS_RECEIVE,
@@ -28,29 +27,10 @@ export const fetchResetOptionsError = ( error ) => ( {
 	error,
 } );
 
-const fromApi = ( data ) => ( [
-	{
-		email: data.primary_email,
-		sms: data.primary_sms,
-	},
-	{
-		email: data.secondary_email,
-		sms: data.secondary_sms,
-	},
-] );
-
-export const fetchResetOptions = ( userData ) => ( dispatch ) => {
-	dispatch( {
-		type: ACCOUNT_RECOVERY_RESET_OPTIONS_REQUEST,
-	} );
-
-	return wpcom.req.get( {
-		body: userData,
-		apiNamespace: 'wpcom/v2',
-		path: '/account-recovery/lookup',
-	} ).then( data => dispatch( fetchResetOptionsSuccess( fromApi( data ) ) ) )
-	.catch( error => dispatch( fetchResetOptionsError( error ) ) );
-};
+export const fetchResetOptions = ( userData ) => ( {
+	type: ACCOUNT_RECOVERY_RESET_OPTIONS_REQUEST,
+	userData,
+} );
 
 export const fetchResetOptionsByLogin = ( user ) => fetchResetOptions( { user } );
